move patterns tests to own module
